🧹 refactor python resource
