+++ providers/os/resources/languages/python/requirements/requirestxt_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) Mondoo, Inc.
 // SPDX-License-Identifier: BUSL-1.1
 
-package python
+package requirements
 
 import (
 	"strings"
