+++ providers/os/resources/languages/python/testdata/rhel.json
